openllama-7b/requirements.txt
requests
python-dotenv
panel
jupyter_bokeh
jupyterlab

